+++ packages/data_taster/spec/data_taster/collection_spec.rb
@@ -20,7 +20,7 @@ RSpec.describe DataTaster::Collection do
     result = DataTaster::Collection.new("projects").assemble
 
     expect(result[:select]).to eq(
-      "INSERT INTO #{dump_db_name}.projects SELECT * FROM #{source_db_name}.projects WHERE 1 = 1"
+      "SELECT * FROM #{source_db_name}.projects WHERE 1 = 1"
     )
   end